+++ Akinita/src/main/java/Akinita/project/Akinita/Entities/Properties/House.java
@@ -49,7 +49,7 @@ public class House extends Property implements BuildingFees, ConstructionDate {
 
     public House() {}
 
-    public House(int id, String estateName, String location, int price, String description, Owner owner, Renter renter, int squareMeter, Boolean availability, String visibility, Date constructionDate, boolean buildingFees) {
+    public House(int id, String estateName, String location, Double price, String description, Owner owner, Renter renter, int squareMeter, Boolean availability, String visibility, Date constructionDate, boolean buildingFees) {
         super(id, estateName, location, price, description, owner, renter, availability, visibility, squareMeter);
         this.constructionDate = constructionDate;
         this.buildingFees = buildingFees;
